Add files via upload
A website was built which can be used to monitor the persons in and out of IOCL dept. as part of security measure. This also
provides the service to track the movement of persons. The same can be monitored via a dashboard by Security Admin.
Technology used: Front-end: HTML, CSS, JavaScript; Backend: PHP, MySQL.